Add support for IT8951 based displays
I've tested this with a 6 inch display that uses the IT8951 controller
board with both VNC and terminal and it seems to work well. I haven't
tested with a 9 inch display but there isn't anything display specific
in the driver (that I'm aware of) so it should just work.

This implantation is based off the WaveShare sample code in
https://github.com/waveshare/IT8951. There are some idiosyncrasies in
that code that I don't understand and aren't documented but appear
necessary. For example, manually controlling the CS pin instead of
letting the SPI driver code handle it...

Issue #25