Optimize transfers for IT8951 displays
Previously the driver was using 8 bits-per-pixel for SPI transfers.
This is wasteful sine the actual display only supports 16 different gray
levels.

This updates the code to use a packed format that uses 4 bits-per-pixel.

There is the potential for further optimization when rendering text (or
other pure black and white images) to use 2 bits-per-pixel. However,
when I tried this the fast display mode did not properly clear pixels
between blits.

Issue #25